Move cosign/pkg/providers to sigstore/pkg/oauth/providers
